pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = [
    "requests==2.31.0",
    "pillow",
    "python-dotenv>=1.0.1",
    "tqdm",
    "gradio==4.7.1",
    "novelai-python>=0.4.0",
    "rtoml>=0.10.0",
    "pydantic-settings>=2.2.1",
]
requires-python = ">=3.8"
